+++ src/aerospikeDownloadUrls.go
@@ -10,8 +10,15 @@ var communityUrls = []string{
 	"https://download.aerospike.com/artifacts/aerospike-server-community/",
 }
 
+var federalUrls = []string{
+	"https://artifacts.aerospike.com/aerospike-server-federal/",
+	"https://download.aerospike.com/artifacts/aerospike-server-federal/",
+}
+
 var promUrl = "https://download.aerospike.com/artifacts/aerospike-prometheus-exporter/"
 
 var enterpriseUrl = enterpriseUrls[1]
 
 var communityUrl = communityUrls[1]
+
+var federalUrl = federalUrls[1]